+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-3947] `UserCache.get_by_user_name` return a user by the user_name
+
 ## [4.25.0] 2021-07-20
 
 ### Added